chore : MenuGroup 패키지 생성 및 변경 리팩터링
