--- src/main/java/kitchenpos/menu/domain/MenuGroupRepository.java
@@ -1,6 +1,0 @@
-package kitchenpos.menu.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
-}
