+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -11,7 +11,7 @@ import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
-import kitchenpos.domain.Price;
+import kitchenpos.common.domain.Price;
 import kitchenpos.menu.dto.MenuProductResponse;
 import kitchenpos.product.domain.Product;
