--- src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -1,40 +1,0 @@
-package kitchenpos.menu.ui;
-
-import java.net.URI;
-import java.util.List;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import kitchenpos.menu.application.MenuGroupService;
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.dto.MenuGroupRequest;
-import kitchenpos.menu.dto.MenuGroupResponse;
-
-@RequestMapping("/api/menu-groups")
-@RestController
-public class MenuGroupRestController {
-	private final MenuGroupService menuGroupService;
-
-	public MenuGroupRestController(final MenuGroupService menuGroupService) {
-		this.menuGroupService = menuGroupService;
-	}
-
-	@PostMapping
-	public ResponseEntity<MenuGroupResponse> create(@RequestBody final MenuGroupRequest menuGroupRequest) {
-		final MenuGroup created = menuGroupService.create(menuGroupRequest);
-		final URI uri = URI.create("/api/menu-groups/" + created.getId());
-		return ResponseEntity.created(uri)
-			.body(created.toResDto());
-	}
-
-	@GetMapping
-	public ResponseEntity<List<MenuGroupResponse>> list() {
-		return ResponseEntity.ok()
-			.body(MenuGroupResponse.ofList(menuGroupService.list()));
-	}
-}
