+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -1,0 +1,29 @@
+package kitchenpos.menugroup.application;
+
+import java.util.List;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
+import kitchenpos.menugroup.dto.MenuGroupRequest;
+
+@Service
+@Transactional(readOnly = true)
+public class MenuGroupService {
+	private final MenuGroupRepository menuGroupRepository;
+
+	public MenuGroupService(final MenuGroupRepository menuGroupRepository) {
+		this.menuGroupRepository = menuGroupRepository;
+	}
+
+	@Transactional
+	public MenuGroup create(final MenuGroupRequest menuGroup) {
+		return menuGroupRepository.save(menuGroup.toEntity());
+	}
+
+	public List<MenuGroup> list() {
+		return menuGroupRepository.findAll();
+	}
+}
