+++ src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.product.dto;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import kitchenpos.domain.Price;
+import kitchenpos.common.domain.Price;
 import kitchenpos.product.domain.Product;
 
 public class ProductResponse {
