--- src/main/java/kitchenpos/tableGroup/domain/TableGroupRepository.java
@@ -1,6 +1,0 @@
-package kitchenpos.tableGroup.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
-}
