--- src/main/java/kitchenpos/tableGroup/exception/TableException.java
@@ -1,10 +1,0 @@
-package kitchenpos.tableGroup.exception;
-
-import kitchenpos.common.BusinessException;
-import kitchenpos.common.ErrorCode;
-
-public class TableException extends BusinessException {
-	public TableException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
