+++ src/main/java/kitchenpos/tablegroup/domain/TableGroupRepository.java
@@ -1,0 +1,6 @@
+package kitchenpos.tablegroup.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
+}
