+++ src/main/java/kitchenpos/tablegroup/exception/TableException.java
@@ -1,0 +1,10 @@
+package kitchenpos.tablegroup.exception;
+
+import kitchenpos.common.BusinessException;
+import kitchenpos.common.ErrorCode;
+
+public class TableException extends BusinessException {
+	public TableException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
