+++ src/test/java/kitchenpos/menu/domain/MenuProductTest.java
@@ -1,12 +1,10 @@
 package kitchenpos.menu.domain;
 
-import static org.assertj.core.api.Assertions.*;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import kitchenpos.domain.Price;
+import kitchenpos.common.domain.Price;
 import kitchenpos.product.domain.Product;
 
 @DisplayName("메뉴 상품 : 단위 테스트")
