--- src/test/java/kitchenpos/menu/exception/MenuException.java
@@ -1,10 +1,0 @@
-package kitchenpos.menu.exception;
-
-import kitchenpos.common.BusinessException;
-import kitchenpos.common.ErrorCode;
-
-public class MenuException extends BusinessException {
-	public MenuException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
